Allow device selection to run ADB command on

+++ src/commands.ts
@@ -14,6 +14,25 @@ const warnAboutMissingAppPackageName = (context: vscode.ExtensionContext): boole
   return false;
 };
 
+const chooseDeviceToRunCommandOn = async (context: vscode.ExtensionContext): Promise<string> => {
+  let activeDevices = (await executeCommand(`adb devices`) as string)
+    .split('\n')
+    .filter((_, index) => index > 0)
+    .map((line) => line.split('\t')[0].trim());
+  if (activeDevices.length === 0) { return ''; }
+  if (activeDevices.length === 1) { return activeDevices[0]; }
+
+  let userSelectedDevice = await vscode.window.showQuickPick(
+    activeDevices,
+    {
+      title: 'Choose target device',
+      canPickMany: false,
+    }
+  );
+
+  return userSelectedDevice ?? '';
+};
+
 const tryExecuteCommands = async (
   commands: Array<string>, successMessage: string, errorMessage: string) => {
   try {
@@ -26,16 +45,16 @@ const tryExecuteCommands = async (
   }
 };
 
-const buildClearDataCommand = (packageName: string): string => {
-  return `adb shell pm clear ${packageName}`;
+const buildClearDataCommand = (packageName: string, targetDevice: string): string => {
+  return `adb -s ${targetDevice} shell pm clear ${packageName}`;
 };
 
-const buildKillCommand = (packageName: string): string => {
-  return `adb shell am force-stop ${packageName}`;
+const buildKillCommand = (packageName: string, targetDevice: string): string => {
+  return `adb -s ${targetDevice} shell am force-stop ${packageName}`;
 };
 
-const buildStartCommand = (packageName: string): string => {
-  return `adb shell monkey - p ${packageName} -c android.intent.category.LAUNCHER 1`;
+const buildStartCommand = (packageName: string, targetDevice: string): string => {
+  return `adb -s ${targetDevice} shell monkey - p ${packageName} -c android.intent.category.LAUNCHER 1`;
 };
 
 export const getAppPackageName = async (context: vscode.ExtensionContext) => {
@@ -60,7 +79,13 @@ export const uninstallApp = async (context: vscode.ExtensionContext) => {
   if (warnAboutMissingAppPackageName(context)) { return; }
   let currentPackageName = getCurrentPackageName(context.workspaceState);
 
-  await tryExecuteCommands([`adb uninstall ${currentPackageName} `],
+  let targetDevice = await chooseDeviceToRunCommandOn(context);
+  if (targetDevice.length === 0) {
+    vscode.window.showErrorMessage('Cannot choose target device to run command on');
+    return;
+  }
+
+  await tryExecuteCommands([`adb -s ${targetDevice} uninstall ${currentPackageName} `],
     `Uninstalled ${currentPackageName} successfuly`,
     `Failed to uninstall ${currentPackageName}`);
 };
@@ -69,7 +94,13 @@ export const killApp = async (context: vscode.ExtensionContext) => {
   if (warnAboutMissingAppPackageName(context)) { return; }
   let currentPackageName = getCurrentPackageName(context.workspaceState);
 
-  await tryExecuteCommands([buildKillCommand(currentPackageName)],
+  let targetDevice = await chooseDeviceToRunCommandOn(context);
+  if (targetDevice.length === 0) {
+    vscode.window.showErrorMessage('Cannot choose target device to run command on');
+    return;
+  }
+
+  await tryExecuteCommands([buildKillCommand(currentPackageName, targetDevice)],
     `Killed ${currentPackageName} successfuly`,
     `Failed to kill ${currentPackageName}`);
 };
@@ -78,7 +109,13 @@ export const startApp = async (context: vscode.ExtensionContext) => {
   if (warnAboutMissingAppPackageName(context)) { return; }
   let currentPackageName = getCurrentPackageName(context.workspaceState);
 
-  await tryExecuteCommands([buildStartCommand(currentPackageName)],
+  let targetDevice = await chooseDeviceToRunCommandOn(context);
+  if (targetDevice.length === 0) {
+    vscode.window.showErrorMessage('Cannot choose target device to run command on');
+    return;
+  }
+
+  await tryExecuteCommands([buildStartCommand(currentPackageName, targetDevice)],
     `Started ${currentPackageName} successfuly`,
     `Failed to start ${currentPackageName}`);
 };
@@ -87,9 +124,15 @@ export const restartApp = async (context: vscode.ExtensionContext) => {
   if (warnAboutMissingAppPackageName(context)) { return; }
   let currentPackageName = getCurrentPackageName(context.workspaceState);
 
+  let targetDevice = await chooseDeviceToRunCommandOn(context);
+  if (targetDevice.length === 0) {
+    vscode.window.showErrorMessage('Cannot choose target device to run command on');
+    return;
+  }
+
   await tryExecuteCommands([
-    buildKillCommand(currentPackageName),
-    buildStartCommand(currentPackageName),
+    buildKillCommand(currentPackageName, targetDevice),
+    buildStartCommand(currentPackageName, targetDevice),
   ],
     `Restarted ${currentPackageName} successfuly`,
     `Failed to restart ${currentPackageName}`);
@@ -99,7 +142,13 @@ export const clearAppData = async (context: vscode.ExtensionContext) => {
   if (warnAboutMissingAppPackageName(context)) { return; }
   let currentPackageName = getCurrentPackageName(context.workspaceState);
 
-  await tryExecuteCommands([buildClearDataCommand(currentPackageName)],
+  let targetDevice = await chooseDeviceToRunCommandOn(context);
+  if (targetDevice.length === 0) {
+    vscode.window.showErrorMessage('Cannot choose target device to run command on');
+    return;
+  }
+
+  await tryExecuteCommands([buildClearDataCommand(currentPackageName, targetDevice)],
     `Cleared ${currentPackageName} data successfuly`,
     `Failed to clear ${currentPackageName} data`);
 };
@@ -108,10 +157,16 @@ export const clearAppDataAndRestart = async (context: vscode.ExtensionContext) =
   if (warnAboutMissingAppPackageName(context)) { return; }
   let currentPackageName = getCurrentPackageName(context.workspaceState);
 
+  let targetDevice = await chooseDeviceToRunCommandOn(context);
+  if (targetDevice.length === 0) {
+    vscode.window.showErrorMessage('Cannot choose target device to run command on');
+    return;
+  }
+
   await tryExecuteCommands([
-    buildClearDataCommand(currentPackageName),
-    buildKillCommand(currentPackageName),
-    buildStartCommand(currentPackageName),
+    buildClearDataCommand(currentPackageName, targetDevice),
+    buildKillCommand(currentPackageName, targetDevice),
+    buildStartCommand(currentPackageName, targetDevice),
   ],
     `Cleared ${currentPackageName} data and restarted successfuly`,
     `Failed to clear ${currentPackageName} data and restart`);
@@ -121,13 +176,19 @@ export const revokeAppPermissions = async (context: vscode.ExtensionContext) => 
   if (warnAboutMissingAppPackageName(context)) { return; }
   let currentPackageName = getCurrentPackageName(context.workspaceState);
 
-  let grantedPermissions = (await executeCommand(`adb shell dumpsys package ${currentPackageName}`) as string)
+  let targetDevice = await chooseDeviceToRunCommandOn(context);
+  if (targetDevice.length === 0) {
+    vscode.window.showErrorMessage('Cannot choose target device to run command on');
+    return;
+  }
+
+  let grantedPermissions = (await executeCommand(`adb -s ${targetDevice} shell dumpsys package ${currentPackageName}`) as string)
     .split('\n')
     .filter((line) => line.indexOf('permission') >= 0 && line.indexOf('granted=true') >= 0)
     .map((line) => line.split(':')[0].trim());
 
   let revokeCommands = grantedPermissions.map(
-    (permission) => `adb shell pm revoke ${currentPackageName} ${permission}`);
+    (permission) => `adb -s ${targetDevice} shell pm revoke ${currentPackageName} ${permission}`);
 
   let failedRevokes = 0;
   for (let command of revokeCommands) {
